requirements.txt
# Core Dependencies
numpy>=1.21.0
pyyaml>=6.0
python-dotenv>=0.19.0
pydantic>=1.9.0
loguru>=0.6.0
sqlalchemy>=2.0.0
alembic>=1.10.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.0  # For efficient similarity search

# Audio Processing
sounddevice>=0.4.4
webrtcvad>=2.0.10
pyaudio>=0.2.12

# LLM Integration
llama-cpp-python>=0.1.78
openai-whisper>=20231117

# Utilities
tqdm>=4.64.0
python-dateutil>=2.8.2

# Development
pytest>=7.0.0
black>=22.3.0
mypy>=0.961
flake8>=4.0.1
